Extract content in separate files
